+++ changelog.md
@@ -24,6 +24,8 @@ Codion Change Log
 - Attribute.DataType.validateType() renamed validate().
 - Attribute caption specified via builder now takes precedence over resource bundle caption.
 - Entity caption specified via builder now takes precedence over resource bundle caption.
+### is.codion.framework.db.core
+- EntityConnectionProvider.Builder.domain(Domain) added.
 ### is.codion.framework.model
 - EntityApplicationModel, EntityModel, EntityEditModel and EntityTableModel, store(Preferences) and restore(Preferences) added.
 - AbstractEntityTableModel now stores sorting state in preferences.